#307 update rpm conf

+++ packages/var-rpm.sh
@@ -9,7 +9,12 @@ fi
 chkconfig --add $bin_name && chkconfig $bin_name on
 "
 
-rpm_postun='if [ `find /opt/echoes-alert/probe -type f | wc -l` -eq 0 ]
+rpm_postun='cd ${PACKAGE_TOP_DIR}
+if [ -d sbin ]
+then
+    rm -rf sbin
+fi
+if [ `find /opt/echoes-alert/probe -type f | wc -l` -eq 0 ]
 then
     rm -rf /opt/echoes-alert/probe
     if [ `find /opt/echoes-alert -type f | wc -l` -eq 0 ]
@@ -20,6 +25,10 @@ fi'
 
 rpm_preun="service $bin_name stop
 cd ${PACKAGE_TOP_DIR}
+if [ -d etc ]
+then
+    rm -rf etc
+fi
 if [ -d lib ]
 then
     rm -rf lib
